+++ bootware.sh
@@ -79,6 +79,15 @@ SUBCOMMANDS:
     uninstall        Remove Bootware files
     update           Update Bootware to latest version
 
+ENVIRONMENT VARIABLES:
+    BOOTWARE_CONFIG     Set the configuration file path
+    BOOTWARE_NOPASSWD   Assume passwordless sudo
+    BOOTWARE_NOSETUP    Skip Ansible install and system setup
+    BOOTWARE_PLAYBOOK   Set Ansible playbook name
+    BOOTWARE_SKIP       Set skip tags for Ansible roles
+    BOOTWARE_TAGS       Set tags for Ansible roles
+    BOOTWARE_URL        Set location of Ansible repository
+
 See 'bootware <subcommand> --help' for more information on a specific command.
 EOF
       ;;