+++ CHANGELOG.md
@@ -9,6 +9,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Deno autocompletion support for Windows.
+- Environment variable documenation.
 - Java, Julia, Parquet, Vagrant, Virtualbox, and Xsv roles.
 
 ## 0.3.4 - 2021-10-20